@@ -7,11 +7,11 @@
     "GPL-2.0-or-later"
   ],
   "require": {
-    "typo3/cms-backend": "^11.4 || ^12.4",
-    "typo3/cms-fluid": "^11.4 || ^12.4",
+    "typo3/cms-backend": "^12.4 || ^13.4",
+    "typo3/cms-fluid": "^12.4 || ^13.4",
     "helhum/typo3-console": "^7.0.0 || ^8.0",
     "zumba/json-serializer": "^3.0",
-    "symfony/process": "^5.0 || ^6.0"
+    "symfony/process": "^6.0 || ^7.0"
   },
   "extra": {
     "typo3/cms": {
