@@ -17,15 +17,15 @@ use Zumba\JsonSerializer\JsonSerializer;
 
 class Job
 {
-    public const STATUS_WATING = 0;
-    public const STATUS_RUNNING = 1;
-    public const STATUS_FINISHED = 2;
-    public const STATUS_KILLED = 3;
-    public const STATUS_FAILED = 4;
+    public const int STATUS_WAITING = 0;
+    public const int STATUS_RUNNING = 1;
+    public const int STATUS_FINISHED = 2;
+    public const int STATUS_KILLED = 3;
+    public const int STATUS_FAILED = 4;
     // can be killed after 10m running
     public const KILLABLE_TIMELIMIT = 600;
 
-    protected $status = self::STATUS_WATING;
+    protected int $status = self::STATUS_WAITING;
     protected Configuration $configuration;
     protected \DateTime $startTime;
     protected \DateTime $endTime;
@@ -126,7 +126,7 @@ class Job
     {
         return
             $this->status === self::STATUS_RUNNING && $this->getExecutionTime() > self::KILLABLE_TIMELIMIT ||
-            $this->status === self::STATUS_WATING
+            $this->status === self::STATUS_WAITING
         ;
     }
 
